refactor(tests): split monolithic repository spec files into dedicated per-repository specs
